requirements.txt
google-genai
beautifulsoup4
requests
html2text
python-dotenv
playwright
langgraph
